requirements.txt
fpdf2==2.8.2
pillow==10.4.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
term_image==0.7.2
